requirements.txt
scrapy
requests
scrapy-playwright
psycopg2-binary
python-dotenv
python-dateutil>=2.8